chore: refactored terminal features
