+++ src/app/domains/terminal/models/autocomplete-result.model.ts
@@ -1,0 +1,6 @@
+import type { LsEntry } from './ls-entry.model';
+
+export interface AutocompleteResult {
+  newCommand: string;
+  suggestions: LsEntry[];
+}
