+++ src/app/domains/terminal/models/file-prefix.type.ts
@@ -1,0 +1,1 @@
+export type FilePrefix = 'dir' | 'arc' | 'bin' | 'lnk' | 'txt' | 'img' | 'vid' | 'aud' | 'unk';
