+++ src/app/domains/terminal/models/ls-entry.model.ts
@@ -1,0 +1,7 @@
+import type { FilePrefix } from './file-prefix.type';
+
+export interface LsEntry {
+  prefix: FilePrefix;
+  name: string;
+  coloredName: string;
+}
