+++ src/app/domains/terminal/models/terminal-line.model.ts
@@ -1,0 +1,11 @@
+import type { SafeHtml } from '@angular/platform-browser';
+import type { LsEntry } from './ls-entry.model';
+
+export type TerminalLineType = 'command' | 'output' | 'error' | 'info' | 'ls';
+
+export interface TerminalLine {
+  type: TerminalLineType;
+  content: SafeHtml | string;
+  timestamp: Date;
+  lsEntries?: LsEntry[];
+}
