+++ src/app/domains/terminal/ui/terminal-panel/terminal.component.css
