+++ src/app/domains/terminal/utils/terminal-autocomplete.utils.ts
@@ -1,49 +1,41 @@
-import { FilePrefix, LsEntry } from './terminal-types';
+import type { AutocompleteResult } from '../models/autocomplete-result.model';
+import type { FilePrefix } from '../models/file-prefix.type';
+import type { LsEntry } from '../models/ls-entry.model';
 
-export interface AutocompleteResult {
-  newCommand: string;
-  suggestions: LsEntry[];
-}
-
-export class TerminalAutocompleteHelper {
+export class TerminalAutocomplete {
   private lastLsEntries: LsEntry[] = [];
 
-  updateCache(lsOutput: string) {
+  updateCache(lsOutput: string): LsEntry[] {
     this.lastLsEntries = this.parseLs(lsOutput);
     return this.lastLsEntries;
   }
 
-  getCache() {
+  getCache(): LsEntry[] {
     return this.lastLsEntries;
   }
 
-  /**
-   * Returns matching suggestions for the current input string.
-   * Logic: Finds the last token of the input and returns all entries starting with that token.
-   * If the input ends in a space, or is empty, or no matches found, returns empty list.
-   */
   getSuggestions(input: string): LsEntry[] {
-    if (this.lastLsEntries.length === 0) return [];
-
-    // If input is empty or just whitespace, don't suggest anything yet (or could suggest all? usually none)
-    if (!input || !input.trim()) return [];
-
-    const raw = input;
-    // If the input ends with a space, we are ready for a new token, so logically we *could* show all files,
-    // but typically shells wait for at least one char or a tab press. The user requested:
-    // "tooltip of the autocomplete neve rresets like it always stay there, (if the command is not here anymore it should disappear)"
-    // So we should only show if there is a partial match on the *current* token.
-    if (/\s$/.test(raw)) {
+    if (this.lastLsEntries.length === 0) {
       return [];
     }
 
-    const trimmed = raw.trim();
+    if (!input || !input.trim()) {
+      return [];
+    }
+
+    if (/\s$/.test(input)) {
+      return [];
+    }
+
+    const trimmed = input.trim();
     const tokens = trimmed.split(/\s+/);
     const lastToken = tokens[tokens.length - 1] || '';
 
-    if (!lastToken) return [];
+    if (!lastToken) {
+      return [];
+    }
 
-    return this.lastLsEntries.filter((e) => e.name.startsWith(lastToken));
+    return this.lastLsEntries.filter((entry) => entry.name.startsWith(lastToken));
   }
 
   handleTab(currentCommand: string): AutocompleteResult {
@@ -52,61 +44,97 @@ export class TerminalAutocompleteHelper {
       suggestions: [],
     };
 
-    if (this.lastLsEntries.length === 0) return result;
+    if (this.lastLsEntries.length === 0) {
+      return result;
+    }
 
     const raw = currentCommand;
     const hasTrailingSpace = /\s$/.test(raw);
     const trimmed = raw.trim();
 
-    if (!trimmed) return result;
+    if (!trimmed) {
+      return result;
+    }
 
     const tokens = trimmed.split(/\s+/);
 
     if (tokens.length === 1 && !hasTrailingSpace) {
-      result.newCommand = raw + ' ';
+      result.newCommand = `${raw} `;
       return result;
     }
 
     const lastToken = hasTrailingSpace ? '' : (tokens[tokens.length - 1] ?? '');
+    const candidates = this.lastLsEntries.filter((entry) => entry.name.startsWith(lastToken));
 
-    const candidates = this.lastLsEntries.filter((e) => e.name.startsWith(lastToken));
-
-    if (candidates.length === 0) return result;
+    if (candidates.length === 0) {
+      return result;
+    }
 
     const prefixText = hasTrailingSpace
       ? raw
       : raw.replace(new RegExp(`${this.escapeRegex(lastToken)}$`), '');
 
     if (candidates.length === 1) {
-      result.newCommand = prefixText + candidates[0].name + ' ';
+      result.newCommand = `${prefixText}${candidates[0].name} `;
       return result;
     }
 
-    const common = this.commonPrefix(candidates.map((c) => c.name));
+    const common = this.commonPrefix(candidates.map((candidate) => candidate.name));
+
     if (common.length > lastToken.length) {
-      result.newCommand = prefixText + common;
+      result.newCommand = `${prefixText}${common}`;
       return result;
     }
 
-    // Multiple candidates, no common prefix extension -> show suggestions
     result.suggestions = candidates;
     return result;
   }
 
+  getClassForPrefix(prefix: FilePrefix): string {
+    switch (prefix) {
+      case 'dir':
+        return 'ft-dir';
+      case 'arc':
+        return 'ft-arc';
+      case 'bin':
+        return 'ft-bin';
+      case 'lnk':
+        return 'ft-lnk';
+      case 'txt':
+        return 'ft-txt';
+      case 'img':
+        return 'ft-img';
+      case 'vid':
+        return 'ft-vid';
+      case 'aud':
+        return 'ft-aud';
+      default:
+        return 'ft-unk';
+    }
+  }
+
   private commonPrefix(items: string[]): string {
-    if (items.length === 0) return '';
+    if (items.length === 0) {
+      return '';
+    }
+
     let prefix = items[0];
+
     for (let i = 1; i < items.length; i++) {
       while (!items[i].startsWith(prefix)) {
         prefix = prefix.slice(0, -1);
-        if (!prefix) return '';
+
+        if (!prefix) {
+          return '';
+        }
       }
     }
+
     return prefix;
   }
 
-  private escapeRegex(s: string): string {
-    return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  private escapeRegex(value: string): string {
+    return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
   }
 
   private parseLs(lsOutput: string): LsEntry[] {
@@ -118,9 +146,12 @@ export class TerminalAutocompleteHelper {
       .map((item) => item.trim())
       .filter((item) => item && item !== 'total' && !/^\d+$/.test(item));
 
-    const out: LsEntry[] = [];
+    const output: LsEntry[] = [];
+
     for (const name of items) {
-      if (name === '.' || name === '..') continue;
+      if (name === '.' || name === '..') {
+        continue;
+      }
 
       let prefix: FilePrefix = typeMap.get(name) || 'unk';
 
@@ -129,9 +160,15 @@ export class TerminalAutocompleteHelper {
       }
 
       const coloredName = coloredMap.get(name) || name;
-      out.push({ prefix, name, coloredName });
+
+      output.push({
+        prefix,
+        name,
+        coloredName,
+      });
     }
-    return out;
+
+    return output;
   }
 
   private detectFileTypeFromName(name: string): FilePrefix {
@@ -158,9 +195,11 @@ export class TerminalAutocompleteHelper {
         'assets',
         'electron',
       ];
+
       if (commonDirs.includes(name.toLowerCase())) {
         return 'dir';
       }
+
       return 'unk';
     }
 
@@ -225,67 +264,64 @@ export class TerminalAutocompleteHelper {
     const typeMap = new Map<string, FilePrefix>();
     const ansiPattern = /((?:\x1B\[[0-9;]*m)+)([^\s\x1B]+)((?:\x1B\[[0-9;]*m)*)/g;
 
-    let match;
+    let match: RegExpExecArray | null;
+
     while ((match = ansiPattern.exec(lsOutput)) !== null) {
       const ansiCodes = match[1] || '';
       const text = match[2];
       const suffix = match[3] || '';
 
-      if (text === 'total' || /^\d+$/.test(text)) continue;
-      if (text === '.' || text === '..') continue;
+      if (text === 'total' || /^\d+$/.test(text) || text === '.' || text === '..') {
+        continue;
+      }
 
       const coloredSegment = ansiCodes + text + suffix;
       coloredMap.set(text, coloredSegment);
 
       const type = this.detectTypeFromAnsiCode(ansiCodes);
+
       if (type !== 'unk') {
         typeMap.set(text, type);
       }
     }
+
     return { coloredMap, typeMap };
   }
 
   private detectTypeFromAnsiCode(ansiCode: string): FilePrefix {
     const match = ansiCode.match(/\[([0-9;]+)m/);
-    if (!match) return 'unk';
 
-    const params = match[1].split(';').map((p) => parseInt(p, 10));
+    if (!match) {
+      return 'unk';
+    }
+
+    const params = match[1].split(';').map((part) => parseInt(part, 10));
 
     for (const param of params) {
-      if (param === 34 || param === 94) return 'dir';
-      if (param === 32 || param === 92) return 'bin';
-      if (param === 36 || param === 96) return 'lnk';
-      if (param === 31 || param === 91) return 'arc';
-      if (param === 35 || param === 95) return 'img';
-      if (param === 33 || param === 93) return 'unk';
+      if (param === 34 || param === 94) {
+        return 'dir';
+      }
+      if (param === 32 || param === 92) {
+        return 'bin';
+      }
+      if (param === 36 || param === 96) {
+        return 'lnk';
+      }
+      if (param === 31 || param === 91) {
+        return 'arc';
+      }
+      if (param === 35 || param === 95) {
+        return 'img';
+      }
+      if (param === 33 || param === 93) {
+        return 'unk';
+      }
     }
+
     return 'unk';
   }
 
   private stripAnsi(text: string): string {
     return text.replace(/\x1B\[[0-9;]*[a-zA-Z]/g, '');
-  }
-
-  public getClassForPrefix(prefix: FilePrefix): string {
-    switch (prefix) {
-      case 'dir':
-        return 'ft-dir';
-      case 'arc':
-        return 'ft-arc';
-      case 'bin':
-        return 'ft-bin';
-      case 'lnk':
-        return 'ft-lnk';
-      case 'txt':
-        return 'ft-txt';
-      case 'img':
-        return 'ft-img';
-      case 'vid':
-        return 'ft-vid';
-      case 'aud':
-        return 'ft-aud';
-      default:
-        return 'ft-unk';
-    }
   }
 }
