+++ src/app/domains/terminal/utils/terminal-history.utils.ts
@@ -1,15 +1,20 @@
-export class TerminalHistoryHelper {
+export class TerminalHistory {
   private history: string[] = [];
-  private index: number = -1;
+  private index = -1;
 
-  add(command: string) {
-    if (!command.trim()) return;
+  add(command: string): void {
+    if (!command.trim()) {
+      return;
+    }
+
     this.history.push(command);
     this.index = this.history.length;
   }
 
   navigate(direction: 'up' | 'down', currentInput: string): string {
-    if (this.history.length === 0) return currentInput;
+    if (this.history.length === 0) {
+      return currentInput;
+    }
 
     if (direction === 'up') {
       if (this.index > 0) {
@@ -20,19 +25,20 @@ export class TerminalHistoryHelper {
       if (this.index < this.history.length - 1) {
         this.index++;
         return this.history[this.index];
-      } else {
-        this.index = this.history.length;
-        return '';
       }
+
+      this.index = this.history.length;
+      return '';
     }
+
     return currentInput;
   }
 
-  reset() {
+  reset(): void {
     this.index = this.history.length;
   }
 
-  setHistory(commands: string[]) {
+  setHistory(commands: string[]): void {
     this.history = [...commands];
     this.index = this.history.length;
   }
